Fixed type inconsistencies, relabeled some variables.

+++ src/lcexplogit.h
@@ -1,6 +1,6 @@
 #ifndef LCEXPLOGIT_HEADER
 #define LCEXPLOGIT_HEADER
 
-double lcexplogit(double *raw_param, int num_types, int num_covariates, int num_students, double *X, int *nskipped, int *nlisted, double *grad);
+double lcexplogit(double *raw_param, int num_types, int num_covariates, int num_students, double *X, uint16_t *nskipped, uint16_t *nlisted, double *grad);
 
 #endif
